parser/requirements.txt
lxml>=4.1.1
requests>=2.18.4
mysql-connector>=2.0.4
